Add files via upload
Started in on some analysis here. I need to write about it, but I'll do that tomorrow. Not everything is in CSV format yet, because I cheated and used some Excel features, so I still need to figure out how to do this without them. Let me know if you have suggestions.